backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
openai
faiss-cpu
python-dotenv
tiktoken
PyPDF2
